extract OL structs into openlineage.rs
