+++ rust/operator-binary/src/main.rs
@@ -51,6 +51,7 @@ mod config;
 mod connect;
 mod crd;
 mod history;
+mod openlineage;
 mod pod_driver_controller;
 mod product_logging;
 mod spark_k8s_controller;
